Unregister databus in store at activity pause
TodoListManager and TodosActivityStore were registred to DataBus,
but since they were created in the context of the Activity they
should have unregistred in onPause to avoid an exeption in the
following onResume when going back/from the application.

Currently all tasks are lost in this case, so next thing to fix
would be to save data (persistently) in the application context
and restore it to the activity when the app is resumed.